requirement.txt
python-barcode
xpinyin